more nbg oversights + stream ttl oversight + debug enhancement in stream (#148)
* more nbg oversights + stream ttl oversight + debug enhancement in stream

* added missing log + fixed log wrong interpolation

* reverted unnecessary code duplication

+++ server/cmwell-ws/app/logic/CRUDServiceFS.scala
@@ -711,7 +711,7 @@ class CRUDServiceFS @Inject()(tbg: NbgToggler)(implicit ec: ExecutionContext, sy
     })
   }
 
-  def scroll(scrollId: String, scrollTTL: Int, withData: Boolean, nbg: Boolean = false): Future[IterationResults] = {
+  def scroll(scrollId: String, scrollTTL: Long, withData: Boolean, nbg: Boolean = false): Future[IterationResults] = {
 
     val searchResultFuture = ftsService(nbg).scroll(scrollId, scrollTTL)
     val results = withData match {
